+++ lib/beamer/locations/settings.dart
@@ -6,6 +6,10 @@ import 'package:lotti/pages/settings/tags/create_tag_page.dart';
 import 'package:lotti/pages/settings/tags/tag_edit_page.dart';
 import 'package:lotti/pages/settings/tags/tags_page.dart';
 
+import '../../pages/settings/dashboards/create_dashboard_page.dart';
+import '../../pages/settings/dashboards/dashboard_definition_page.dart';
+import '../../pages/settings/dashboards/dashboards_page.dart';
+
 class SettingsLocation extends BeamLocation<BeamState> {
   SettingsLocation(RouteInformation super.routeInformation);
 
@@ -15,32 +19,34 @@ class SettingsLocation extends BeamLocation<BeamState> {
         '/settings/tags',
         '/settings/tags/:tagEntityId',
         '/settings/tags/create/:tagType',
+        '/settings/dashboards',
+        '/settings/dashboards/:dashboardId',
+        '/settings/dashboards/create',
       ];
 
   @override
   List<BeamPage> buildPages(BuildContext context, BeamState state) {
-    final pages = [
+    bool pathContains(String s) => state.uri.path.contains(s);
+    bool pathContainsKey(String s) => state.pathParameters.containsKey(s);
+
+    return [
       const BeamPage(
         key: ValueKey('settings'),
         title: 'Settings',
         type: BeamPageType.noTransition,
         child: SettingsPage(),
       ),
-    ];
 
-    if (state.uri.path.contains('tags')) {
-      pages.add(
+      // Tags
+      if (pathContains('tags'))
         const BeamPage(
           key: ValueKey('settings-tags'),
           child: TagsPage(),
         ),
-      );
-    }
 
-    if (state.uri.path.contains('tags') &&
-        !state.uri.path.contains('create') &&
-        state.pathParameters.containsKey('tagEntityId')) {
-      pages.add(
+      if (pathContains('tags') &&
+          !pathContains('create') &&
+          pathContainsKey('tagEntityId'))
         BeamPage(
           key: ValueKey(
             'settings-tags-${state.pathParameters['tagEntityId']}',
@@ -49,22 +55,40 @@ class SettingsLocation extends BeamLocation<BeamState> {
             tagEntityId: state.pathParameters['tagEntityId']!,
           ),
         ),
-      );
-    }
 
-    if (state.uri.path.contains('tags/create') &&
-        state.pathParameters.containsKey('tagType')) {
-      pages.add(
+      if (pathContains('tags/create') && pathContainsKey('tagType'))
         BeamPage(
           key: ValueKey(
             'settings-tags-create-${state.pathParameters['tagType']}',
           ),
           child: CreateTagPage(tagType: state.pathParameters['tagType']!),
         ),
-      );
-    }
 
-    return pages;
+      // Dashboards
+      if (pathContains('dashboards'))
+        const BeamPage(
+          key: ValueKey('settings-dashboards'),
+          child: DashboardSettingsPage(),
+        ),
+
+      if (pathContains('dashboards') &&
+          !pathContains('create') &&
+          pathContainsKey('dashboardId'))
+        BeamPage(
+          key: ValueKey(
+            'settings-dashboards-${state.pathParameters['dashboardId']}',
+          ),
+          child: EditDashboardPage(
+            dashboardId: state.pathParameters['dashboardId']!,
+          ),
+        ),
+
+      if (pathContains('dashboards/create'))
+        const BeamPage(
+          key: ValueKey('settings-dashboards-create'),
+          child: CreateDashboardPage(),
+        ),
+    ];
   }
 }