+++ lib/widgets/settings/dashboards/dashboard_definition_card.dart
@@ -1,3 +1,4 @@
+import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/get_it.dart';
@@ -5,6 +6,9 @@ import 'package:lotti/routes/router.gr.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+
+import '../../../database/database.dart';
+import '../../../utils/consts.dart';
 
 class DashboardDefinitionCard extends StatelessWidget {
   const DashboardDefinitionCard({
@@ -19,6 +23,18 @@ class DashboardDefinitionCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final id = dashboard.id;
+    void beamToNamed(String path) => context.beamToNamed(path);
+
+    Future<void> onTap() async {
+      final beamerNav =
+          await getIt<JournalDb>().getConfigFlag(enableBeamerNavFlag);
+
+      if (beamerNav) {
+        beamToNamed('/settings/dashboards/$id');
+      } else {
+        await getIt<AppRouter>().push(EditDashboardRoute(dashboardId: id));
+      }
+    }
 
     return Card(
       color: colorConfig().headerBgColor,
@@ -70,8 +86,7 @@ class DashboardDefinitionCard extends StatelessWidget {
             ),
           ],
         ),
-        onTap: () =>
-            getIt<AppRouter>().push(EditDashboardRoute(dashboardId: id)),
+        onTap: onTap,
       ),
     );
   }